refactor metodo obtener monopatines con viajes en cierto año

+++ microservicio-administracion/src/main/java/com/tudai/aw/ms_administracion/model/dto/MonopatinConViajesDto.java
@@ -6,11 +6,11 @@ import lombok.Getter;
 public class MonopatinConViajesDto {
 	private Long id;
 	private Long cantViajes;
-	private Long anio;
+	private Integer anio;
 	
 	public MonopatinConViajesDto() {}
 	
-	public MonopatinConViajesDto(Long id, Long cantViajes, Long anio) {
+	public MonopatinConViajesDto(Long id, Long cantViajes, Integer anio) {
 		super();
 		this.id = id;
 		this.cantViajes = cantViajes;
